projects/rawg_pipeline/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
agate==1.9.1
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
annotated-types==0.7.0
    # via pydantic
attrs==26.1.0
    # via
    #   jsonschema
    #   referencing
babel==2.18.0
    # via agate
certifi==2026.4.22
    # via requests
charset-normalizer==3.4.7
    # via requests
click==8.3.3
    # via
    #   dbt-core
    #   dbt-semantic-interfaces
colorama==0.4.6
    # via dbt-common
daff==1.4.2
    # via dbt-core
dbt-adapters==1.23.0
    # via
    #   dbt-core
    #   dbt-duckdb
dbt-common==1.38.0
    # via
    #   dbt-adapters
    #   dbt-core
    #   dbt-duckdb
dbt-core==1.11.9
    # via dbt-duckdb
dbt-duckdb==1.10.1
    # via -r requirements.in
dbt-extractor==0.6.0
    # via dbt-core
dbt-protos==1.0.491
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
dbt-semantic-interfaces==0.9.0
    # via dbt-core
deepdiff==8.6.2
    # via dbt-common
duckdb==1.5.2
    # via
    #   -r requirements.in
    #   dbt-duckdb
    #   duckdb-engine
duckdb-engine==0.17.0
    # via -r requirements.in
greenlet==3.5.0
    # via sqlalchemy
idna==3.14
    # via requests
importlib-metadata==8.9.0
    # via dbt-semantic-interfaces
iniconfig==2.3.0
    # via pytest
isodate==0.7.2
    # via
    #   agate
    #   dbt-common
jinja2==3.1.6
    # via
    #   dbt-common
    #   dbt-core
    #   dbt-semantic-interfaces
jsonschema==4.26.0
    # via
    #   dbt-common
    #   dbt-core
    #   dbt-semantic-interfaces
jsonschema-specifications==2025.9.1
    # via jsonschema
leather==0.4.1
    # via agate
markupsafe==3.0.3
    # via jinja2
mashumaro[msgpack]==3.14
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
more-itertools==10.8.0
    # via dbt-semantic-interfaces
msgpack==1.1.2
    # via mashumaro
networkx==3.6.1
    # via dbt-core
numpy==2.4.4
    # via pandas
orderly-set==5.5.0
    # via deepdiff
packaging==26.2
    # via
    #   dbt-core
    #   duckdb-engine
    #   pytest
pandas==3.0.3
    # via -r requirements.in
parsedatetime==2.6
    # via agate
pathspec==0.12.1
    # via
    #   dbt-common
    #   dbt-core
pluggy==1.6.0
    # via pytest
protobuf==6.33.6
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
    #   dbt-protos
pyarrow==24.0.0
    # via -r requirements.in
pydantic==2.13.4
    # via
    #   dbt-core
    #   dbt-semantic-interfaces
pydantic-core==2.46.4
    # via pydantic
pygments==2.20.0
    # via pytest
pytest==9.0.3
    # via
    #   -r requirements.in
    #   pytest-mock
pytest-mock==3.15.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   dbt-common
    #   dbt-semantic-interfaces
    #   pandas
python-dotenv==1.2.2
    # via -r requirements.in
python-slugify==8.0.4
    # via agate
pytimeparse==1.1.8
    # via agate
pytz==2026.2
    # via
    #   dbt-adapters
    #   dbt-core
pyyaml==6.0.3
    # via
    #   dbt-core
    #   dbt-semantic-interfaces
    #   responses
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.33.1
    # via
    #   -r requirements.in
    #   dbt-common
    #   dbt-core
    #   responses
    #   snowplow-tracker
responses==0.26.0
    # via -r requirements.in
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
ruff==0.15.12
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
snowplow-tracker==1.1.0
    # via dbt-core
sqlalchemy==2.0.49
    # via
    #   -r requirements.in
    #   duckdb-engine
sqlparse==0.5.5
    # via dbt-core
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.15.0
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
    #   dbt-semantic-interfaces
    #   mashumaro
    #   pydantic
    #   pydantic-core
    #   referencing
    #   snowplow-tracker
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.7.0
    # via
    #   requests
    #   responses
zipp==3.23.1
    # via importlib-metadata
